+++ Chapter.Net.WPF.Converters.Tests/BooleanToBooleanConverter/BooleanToBooleanConverterTests.cs
@@ -4,8 +4,8 @@
 // </copyright>
 // -----------------------------------------------------------------------------------------------------------------
 
-using NUnit.Framework;
 using System;
+using NUnit.Framework;
 
 // ReSharper disable once CheckNamespace
 
@@ -13,68 +13,74 @@ namespace Chapter.Net.WPF.Converters.Tests;
 
 public class BooleanToBooleanConverterTests : SingleAndMultiValueConverterTester<BooleanToBooleanConverter>
 {
-    [TestCase(true, true, true, true)]
-    [TestCase(false, true, true, false)]
-    [TestCase(null, true, true, null)]
-    [TestCase(true, true, false, true)]
-    [TestCase(true, false, false, false)]
-    [TestCase(true, null, false, null)]
-    [TestCase(true, true, "", true)]
-    [TestCase(true, true, 15, true)]
-    [TestCase(true, true, null, true)]
-    [TestCase(true, false, "", false)]
-    [TestCase(true, false, 15, false)]
-    [TestCase(true, false, null, false)]
-    [TestCase(true, null, "", null)]
-    [TestCase(true, null, 15, null)]
-    [TestCase(true, null, null, null)]
-    public void Convert_Called_Converts(bool? trueIs, bool? falseIs, object input, bool? expectation)
+    [TestCase(true, true, null, true, true)]
+    [TestCase(false, true, null, true, false)]
+    [TestCase(null, true, null, true, null)]
+    [TestCase(true, true, null, false, true)]
+    [TestCase(true, false, null, false, false)]
+    [TestCase(true, null, null, false, null)]
+    [TestCase(true, false, true, null, true)]
+    [TestCase(true, false, false, null, false)]
+    [TestCase(true, false, null, null, null)]
+    [TestCase(true, true, null, "", true)]
+    [TestCase(true, false, null, "", false)]
+    [TestCase(true, null, null, "", null)]
+    [TestCase(true, true, null, 15, true)]
+    [TestCase(true, false, null, 15, false)]
+    [TestCase(true, null, null, 15, null)]
+    public void Convert_Called_Converts(bool? trueIs, bool? falseIs, bool? nullIs, object input, bool? expectation)
     {
         _target.TrueIs = trueIs;
         _target.FalseIs = falseIs;
+        _target.NullIs = nullIs;
 
         Convert(input, expectation);
     }
 
-    [TestCase(true, true, true, true)]
-    [TestCase(false, true, true, false)]
-    [TestCase(null, true, true, null)]
-    [TestCase(true, true, false, true)]
-    [TestCase(true, false, false, false)]
-    [TestCase(true, null, false, null)]
-    [TestCase(true, true, "", true)]
-    [TestCase(true, true, 15, true)]
-    [TestCase(true, true, null, true)]
-    [TestCase(true, false, "", false)]
-    [TestCase(true, false, 15, false)]
-    [TestCase(true, false, null, false)]
-    [TestCase(true, null, "", null)]
-    [TestCase(true, null, 15, null)]
-    [TestCase(true, null, null, null)]
-    public void ConvertBack_Called_Converts(bool? trueIs, bool? falseIs, object input, bool? expectation)
+    [TestCase(true, true, null, true, true)]
+    [TestCase(false, true, null, true, false)]
+    [TestCase(null, true, null, true, null)]
+    [TestCase(true, true, null, false, true)]
+    [TestCase(true, false, null, false, false)]
+    [TestCase(true, null, null, false, null)]
+    [TestCase(true, false, true, null, true)]
+    [TestCase(true, false, false, null, false)]
+    [TestCase(true, false, null, null, null)]
+    [TestCase(true, true, null, "", true)]
+    [TestCase(true, false, null, "", false)]
+    [TestCase(true, null, null, "", null)]
+    [TestCase(true, true, null, 15, true)]
+    [TestCase(true, false, null, 15, false)]
+    [TestCase(true, null, null, 15, null)]
+    public void ConvertBack_Called_Converts(bool? trueIs, bool? falseIs, bool? nullIs, object input, bool? expectation)
     {
         _target.TrueIs = trueIs;
         _target.FalseIs = falseIs;
+        _target.NullIs = nullIs;
 
         ConvertBack(input, expectation);
     }
 
-    [TestCase(true, true, true, true, true, true, true)]
-    [TestCase(false, true, true, false, true, true, true)]
-    [TestCase(null, true, true, null, true, true, true)]
-    [TestCase(true, true, true, true, false, false, false)]
-    [TestCase(true, false, true, false, false, false, false)]
-    [TestCase(true, null, true, null, false, false, false)]
-    [TestCase(true, true, true, true, false, true, false)]
-    [TestCase(true, true, false, false, false, true, false)]
-    [TestCase(true, true, null, null, false, true, false)]
-    [TestCase(true, false, null, true, true, "", true)]
-    [TestCase(true, false, null, false, false, "", false)]
-    [TestCase(true, false, null, null, true, "", false)]
-    public void MultiConvert_Called_Converts(bool? trueIs, bool? falseIs, bool? mixedIs, bool? expectation, params object[] input)
+    [TestCase(true, true, null, false, true, true, true, true)]
+    [TestCase(false, true, null, false, false, true, true, true)]
+    [TestCase(null, true, null, false, null, true, true, true)]
+    [TestCase(true, true, null, false, true, false, false, false)]
+    [TestCase(true, false, null, false, false, false, false, false)]
+    [TestCase(true, null, null, false, null, false, false, false)]
+    [TestCase(true, false, true, false, true, null, null, null)]
+    [TestCase(true, false, false, false, false, null, null, null)]
+    [TestCase(true, false, null, false, null, null, null, null)]
+    [TestCase(true, false, null, true, true, false, true, null)]
+    [TestCase(true, false, null, false, false, false, true, null)]
+    [TestCase(true, false, null, null, null, false, true, null)]
+    [TestCase(true, false, null, true, true, true, "", true)]
+    [TestCase(true, false, null, false, false, false, "", false)]
+    [TestCase(true, false, null, null, null, true, "", false)]
+    public void MultiConvert_Called_Converts(bool? trueIs, bool? falseIs, bool? nullIs, bool? mixedIs, bool? expectation, params object[] input)
     {
         _target.TrueIs = trueIs;
         _target.FalseIs = falseIs;
+        _target.NullIs = nullIs;
         _target.MixedIs = mixedIs;
 
         MultiConvert(input, expectation);